+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Moved and renamed target transform utility functions
 - Examples have two levels of headings in the table of content
 - Fix orders of examples in table of content
+- `DiscreteCustomConstraint` validator now expects data frame instead of series
 
 ### Fixed
 - Wrong use of `tolerance` argument in constraints user guide